+++ CuttingEdge/CEPoint2d.swift
@@ -7,9 +7,9 @@
 //
 
 public struct Point2d: Equatable {
-    var X: Int = 0
-    var Y: Int = 0
-    var Z: Int = 0
+    var X: Float = 0
+    var Y: Float = 0
+    var Z: Float = 0
     
     public static func == (left: Point2d, right: Point2d) -> Bool {
         return ( ( left.X == right.X) && (left.Y == right.Y) )
